pyproject.toml
[project]
name = "django-template"
version = "0.1.0"
description = "This template is build to accelerate development in dev and migration to stage and production."
authors = [
    {name = "Mohd Zakir Hussain",email = "mohd.zakir.hussain@outlook.com"}
]
readme = "README.md"
requires-python = ">=3.12,<3.13     "
dependencies = [
    "django (>=5.2.1,<6.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "django-environ (>=0.12.0,<0.13.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "retell-sdk (>=4.32.0,<5.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "requests (>=2.32.3,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.1"
flake8 = "^7.2.0"
pytest = "^8.3.5"

